requirements.txt
keyboard
google-cloud-texttospeech
pyyaml
pygame
setuptools<78.0.1